double statt float

+++ include/JoinQuery.hpp
@@ -67,7 +67,7 @@ class JoinQuery
      * @param orderIds
      * @return all quantities of table lineitem belonging to the orders from the parameter orderIds
      */
-    std::multiset<float>getLineitemQuantities(std::unordered_set<int>orderIds);
+    std::multiset<double>getLineitemQuantities(std::unordered_set<int>orderIds);
    /// Returns line count of given file
    size_t lineCount(std::string rel);
 };
